Check cluster version when creating a cluster
