+++ test/commands/register.test.ts
@@ -10,7 +10,6 @@ import ComponentRegister from '../../src/commands/register';
 import { DockerComposeUtils } from '../../src/common/docker-compose';
 import DockerComposeTemplate from '../../src/common/docker-compose/template';
 import DockerBuildXUtils from '../../src/common/docker/buildx.utils';
-import { KubeClusterUtils } from '../../src/common/utils/kube-cluster.utils';
 import { IF_EXPRESSION_REGEX } from '../../src/dependency-manager/spec/utils/interpolation';
 import { mockArchitectAuth, MOCK_API_HOST, MOCK_REGISTRY_HOST } from '../utils/mocks';
 
@@ -38,9 +37,7 @@ describe('register', function () {
     name: 'architect',
   };
 
-  const min_cluster_version = { 'major': 1, 'minor': 22, 'gitVersion': 'v1.22.0' };
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(ComponentRegister, 'registerComponent', sinon.stub().returns({}))
     .stdout({ print })
@@ -52,7 +49,6 @@ describe('register', function () {
     .it('expect default project path to be ./architect.yml if not provided');
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_architect_account_response)
@@ -79,8 +75,7 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-  .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
-  .stub(AccountUtils, 'isValidAccount', sinon.stub().returns(false))
+    .stub(AccountUtils, 'isValidAccount', sinon.stub().returns(false))
     .stub(DockerBuildXUtils, 'convertToBuildxPlatforms', sinon.stub().returns([]))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
@@ -105,7 +100,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(DockerBuildXUtils, 'convertToBuildxPlatforms', sinon.stub().throws(new Error('Some internal docker build exception')))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
@@ -121,7 +115,6 @@ describe('register', function () {
     .it('register component with architecture flag failed');
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
       .head(/.*/)
@@ -176,7 +169,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_account_response)
@@ -202,7 +194,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_account_response)
@@ -239,7 +230,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_account_response)
@@ -265,7 +255,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_account_response)
@@ -292,7 +281,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
       .reply(200, mock_account_response)
@@ -314,7 +302,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(AccountUtils, 'isValidAccount', sinon.stub().returns(false))
     .nock(MOCK_API_HOST, api => api
       .get('/accounts/examples')
@@ -332,7 +319,6 @@ describe('register', function () {
     .it('rejects with informative error message if account is unavailable');
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
@@ -365,7 +351,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub().throws(new Error('Some internal docker build exception')))
     .nock(MOCK_API_HOST, api => api
       .get(`/accounts/examples`)
@@ -381,7 +366,6 @@ describe('register', function () {
     .it('rejects with the original error message if docker buildx inspect fails');
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
@@ -410,7 +394,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .stub(DockerComposeUtils, 'writeCompose', sinon.stub())
@@ -449,7 +432,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .nock(MOCK_REGISTRY_HOST, api => api
@@ -480,7 +462,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
       .head(/.*/)
@@ -506,7 +487,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
       .head(/.*/)
@@ -532,7 +512,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
       .head(/.*/)
@@ -558,7 +537,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
       .head(/.*/)
@@ -584,7 +562,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .stub(DockerComposeUtils, 'writeCompose', sinon.stub())
@@ -617,7 +594,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .stub(DockerComposeUtils, 'writeCompose', sinon.stub())
@@ -651,7 +627,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .stub(DockerComposeUtils, 'writeCompose', sinon.stub())
@@ -685,7 +660,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(fs, 'move', sinon.stub())
     .stub(DockerBuildXUtils, 'dockerBuildX', sinon.stub())
     .stub(DockerComposeUtils, 'writeCompose', sinon.stub())
@@ -718,7 +692,6 @@ describe('register', function () {
     });
 
   mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
     .stub(AccountUtils, 'isValidAccount', sinon.stub().returns(false))
     .nock(MOCK_REGISTRY_HOST, api => api
       .persist()
@@ -743,37 +716,4 @@ describe('register', function () {
       expect(ctx.stdout).to.contain(`The account name 'invalid-account' was found as part of the component name in your architect.yml file. Either that account does not exist or you do not have permission to access it.`);
       expect(ctx.stdout).to.contain('Successfully registered component');
     });
-  
-  mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns(min_cluster_version))
-    .stub(AccountUtils, 'isValidAccount', sinon.stub().returns(false))
-    .stub(DockerBuildXUtils, 'convertToBuildxPlatforms', sinon.stub().returns([]))
-    .nock(MOCK_API_HOST, api => api
-      .get(`/accounts/examples`)
-      .reply(200, mock_architect_account_response)
-    )
-    .nock(MOCK_REGISTRY_HOST, api => api
-      .persist()
-      .head(/.*/)
-      .reply(200, '', { 'docker-content-digest': 'some-digest' })
-    )
-    .nock(MOCK_API_HOST, api => api
-      .post(/\/accounts\/.*\/components/, (body) => body)
-      .reply(200))
-    .stdout({ print })
-    .stderr({ print })
-    .command(['register', 'examples/hello-world/architect.yml', '-t', '1.0.0', '-a', 'examples'])
-    .it(`register with minimum cluster version`, ctx => {
-      expect(ctx.stdout).to.contain('Successfully registered component');
-    });
-
-  mockArchitectAuth()
-    .stub(KubeClusterUtils, 'getClientVersion', sinon.stub().returns({ 'major': 1, 'minor': 0, 'gitVersion': 'v1.0.0' }))
-    .stdout({ print })
-    .stderr({ print })
-    .command(['register', 'examples/hello-world/architect.yml', '-t', '1.0.0', '-a', 'examples'])
-    .catch(e => {
-      expect(e.message).contains(`Currently, we only support Kubernetes clusters on version ${min_cluster_version.major}.${min_cluster_version.minor} or greater.`);
-    })
-    .it('register with older cluster version fails');
 });
